move LSP*EventCreators to subpackage
