+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/eventsCreator/LSPFreightLinkEnterEventCreator.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.freight.events;
+package org.matsim.contrib.freight.events.eventsCreator;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.events.Event;
@@ -7,6 +7,7 @@ import org.matsim.api.core.v01.population.Activity;
 import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
+import org.matsim.contrib.freight.events.LSPFreightLinkEnterEvent;
 
 /*package-private*/ final class LSPFreightLinkEnterEventCreator implements LSPEventCreator {
