+++ contribs/freight/src/main/java/org/matsim/contrib/freight/events/eventsCreator/LSPTourStartEventCreator.java
@@ -1,4 +1,4 @@
-package org.matsim.contrib.freight.events;
+package org.matsim.contrib.freight.events.eventsCreator;
 
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.events.ActivityEndEvent;
@@ -8,6 +8,7 @@ import org.matsim.api.core.v01.population.Person;
 import org.matsim.contrib.freight.carrier.Carrier;
 import org.matsim.contrib.freight.carrier.FreightConstants;
 import org.matsim.contrib.freight.carrier.ScheduledTour;
+import org.matsim.contrib.freight.events.LSPTourStartEvent;
 
 /*package-private*/  final class LSPTourStartEventCreator implements LSPEventCreator {
